corrections on new regressions, only joint 6 not working good, and pb of CAN BUS not detected

+++ niryo_robot_hardware_stack/common/src/model/single_motor_cmd.cpp
@@ -65,7 +65,7 @@ namespace common {
             ss << DxlCommandTypeEnum(_type).toString();
 
             ss << ": ";
-            ss << "motor " << static_cast<int>(_id) << ": " << _param;
+            ss << "motor " << static_cast<int>(_id) << ": " << static_cast<int>(_param);
 
             return ss.str();
         }
